+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- MP3 ID3v2, OGG, FLAC and MPEG video support
 
 ## 0.3.0 – 2017-08-06
 ### Added